+++ examples/nextjs-sync/.next/static/chunks/596.b14d205d2e5128f5.js
@@ -1,1 +1,0 @@
-"use strict";(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[596],{7596:function(n,u,e){e.r(u),e.d(u,{TalaDBNode:function(){return a}});var a=null}}]);